worker/requirements.txt
Pillow
transformers
torch
pytest
pytest-cov 